Separated routes into different controllers
